[TT-003/004] Further Refactoring [EP, CS]
