+++ TodoTasker/app/build.gradle
@@ -10,7 +10,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.todo_tasker"
+        applicationId "com.backend.todo_tasker"
         minSdkVersion 26
         targetSdkVersion 30
         versionCode 1
