+++ TodoTasker/app/src/androidTest/java/com/backend/todo_tasker/DataBaseReminderInstrumentedTest.kt
@@ -1,7 +1,9 @@
-package com.example.todo_tasker
+package com.backend.todo_tasker
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.backend.todo_tasker.database.DatabaseClass
+import com.backend.todo_tasker.database.Todo
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -14,7 +16,8 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class DataBaseAddInstrumentedTest {
+class DataBaseReminderInstrumentedTest {
+    // Single entry with reminder info
     @Test
     fun SingleAddPass() {
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
@@ -24,12 +27,13 @@ class DataBaseAddInstrumentedTest {
         db.deleteDBEntries(datab)
 
         val time_in_millis = db.dateToMillis(db.getCurrentDate())
-        val test_todo = Todo(6, "Test", time_in_millis, null)
+        val test_todo = Todo(6, "Test", time_in_millis, time_in_millis)
 
         val ret_val = db.addToDb(datab, test_todo)
         assertNotEquals(ret_val, -1)
     }
 
+    // Loads of Entries with reminder info
     @Test
     fun MultiAddPass() {
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
@@ -40,12 +44,13 @@ class DataBaseAddInstrumentedTest {
 
         val time_in_millis = db.dateToMillis(db.getCurrentDate())
         for(i in 0..50) {
-            val test_todo = Todo(i, "Test", time_in_millis, null)
+            val test_todo = Todo(i, "Test", time_in_millis, time_in_millis)
             val ret_val = db.addToDb(datab, test_todo)
             assertNotEquals(ret_val, -1)
         }
     }
 
+    // Loads of Entries that should fail as the UID is the same
     @Test
     fun MultiUIDFail() {
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
@@ -55,12 +60,12 @@ class DataBaseAddInstrumentedTest {
         db.deleteDBEntries(datab)
 
         val time_in_millis = db.dateToMillis(db.getCurrentDate())
-        var test_todo = Todo(1337, "TestThatPasses", time_in_millis, null)
+        var test_todo = Todo(1337, "TestThatPasses", time_in_millis, time_in_millis)
         var ret_val = db.addToDb(datab, test_todo)
         assertNotEquals(ret_val, -1)
 
         for(i in 0..50) {
-            test_todo = Todo(1337, "TestThatFails", time_in_millis, null)
+            test_todo = Todo(1337, "TestThatFails", time_in_millis, time_in_millis)
             ret_val = db.addToDb(datab, test_todo)
             assertEquals(ret_val, -1)
         }
