+++ TodoTasker/app/src/androidTest/java/com/backend/todo_tasker/DataBaseSelectInstrumentedTest.kt
@@ -1,7 +1,9 @@
-package com.example.todo_tasker
+package com.backend.todo_tasker
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.backend.todo_tasker.database.DatabaseClass
+import com.backend.todo_tasker.database.Todo
 
 import org.junit.Test
 import org.junit.runner.RunWith
