+++ TodoTasker/app/src/androidTest/java/com/backend/todo_tasker/UITestExample.kt
@@ -1,4 +1,4 @@
-package com.example.todo_tasker
+package com.backend.todo_tasker
 
 /*
  DISCLAIMER:
@@ -6,20 +6,6 @@ package com.example.todo_tasker
     DO NOT DELETE
 */
 
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.espresso.matcher.ViewMatchers.withText
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.LargeTest
-import org.junit.Assert.*
 /*
 @RunWith(AndroidJUnit4::class)
 @LargeTest
