+++ TodoTasker/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.todo_tasker">
+    package="com.backend.todo_tasker">
 
     <uses-permission android:name="android.permission.WAKE_LOCK" />
     <application
