+++ TodoTasker/app/src/main/java/com/backend/todo_tasker/database/TodoDatabase.kt
@@ -1,0 +1,9 @@
+package com.backend.todo_tasker.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = [Todo::class], version = 1)
+abstract class TodoDatabase : RoomDatabase() {
+    abstract fun todoDao(): TodoDao
+}
